Add linguistic schema and some basic tests
This schema can account for every root, stem, and form in Sanskrit.
This includes prefixes and indeclinables.

The tests provided examine only basic properties of the different
models, but they're sufficient to verify that there are no major bugs.

The setup code can set up only basic models.

The schema has several interlocking parts, and the design of one
component affects the design of several others. This explains, in
part, why this commit is so monolithic.