Add setup code for some models and refactor
The models include the following:

- sandhi rules
- indeclinables
- verb prefixes
- verbal stuff:
  - roots
  - prefixed roots
  - verbs (unmodified only)
- pronouns

Along the way, gut the whole "enum" implementation -- hard-coding IDs
in classes is inelegant and unsafe. Unfortunately, we have to keep the
hard-coded IDs in 'Tag' for polymorphic joins.

Also, refactor all of the schema definitions into one file. Due to the
multitude of names involved, one test case ended up using a compiled
version of a file that no longer existed, which was resolved after
lots of debugging.

Additionally, refactor database code into separate files and merge
'sanskrit.db' with the main package.