Create 'GenderGroup' for stems
Originally, 'Gender' was overloaded to refer to both form genders and
stem genders. Since stems can have multiple genders, 'Gender' held
genders like 'm' and 'f' in addition to 'mf'.

But this makes it difficult to reason about the form genders that are
compatible with a stem gender. Suppose we have a stem with genders
'mn'. The only way to decompose these genders is to read the 'abbr'
field from the gender, iterate over the letters in 'abbr', and do a
reverse lookup for other genders. This imposes a restriction on the
values of 'abbr' that can only be represented as code.

This approach muddles the semantics of the truth of the situtaion: a
stem *has many* genders, and the schema should reflect that.