Remove analyze.Nominal
I think I initially created it because I wasn't sure how the ORM would
react if I used a "hallucinated" record. I no longer believe this is a
cause for concern because SQLAlchemy is fairly good about separating
class and SQL logic and because having a second Nominal class is
confusing.